細かな修正

+++ resources/views/projects/show.blade.php
@@ -8,10 +8,15 @@
     </head>
     <body>
         <div class="user">
-            投稿者：{{ $project->user->name }}
+            <x-user-icon>
+                <img src="{{ $project->user->image_path }}" />
+            </x-user-icon>
+            {{ $project->user->name }}
         </div>
         <div class="edit">
-            <a href="/projects/{{ $project->id }}/edit">投稿を編集</a>
+            @if(Auth::User()->id === $project->user_id)
+                <a href="/projects/{{ $project->id }}/edit">投稿を編集</a>
+            @endif
         </div>
         <div class="show_project">
             <h1>{{ $project->title }}</h1>
@@ -29,11 +34,13 @@
         </div>
         <div class="footer">
         <div class="delete">
-            <form method="POST" action="/projects/{{ $project->id }}" id="form_{{ $project->id }}">
-                @csrf
-                @method("DELETE")
-                <input type="button" onclick="deleteProject({{ $project->id }})" value="投稿を削除" />
-            </form>
+            @if(Auth::user()->id === $project->user_id)
+                <form method="POST" action="/projects/{{ $project->id }}" id="form_{{ $project->id }}">
+                    @csrf
+                    @method("DELETE")
+                    <input type="button" onclick="deleteProject({{ $project->id }})" value="投稿を削除" />
+                </form>
+            @endif
         <div class="deliverables">
             <p><a href="{{ route('apps.create', ['project' => $project->id]) }}">成果物投稿</a></p>
         </div>
